% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squirrel_function.R
\name{squirrel_function}
\alias{squirrel_function}
\title{A Squirrel Function}
\usage{
squirrel_function(love = TRUE)
}
\arguments{
\item{admire}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
squirrel_function()
}
